Extract to a lib.rs and do tests in there
